refactor: 뒤로가기 종료 확인 로직 ExitConfirmationHandler로 분리
