+++ android/app/src/main/java/com/on/turip/ui/compose/main/component/MainApp.kt
@@ -1,6 +1,5 @@
 package com.on.turip.ui.compose.main.component
 
-import androidx.activity.compose.BackHandler
 import androidx.compose.animation.ContentTransform
 import androidx.compose.animation.core.animateDpAsState
 import androidx.compose.animation.core.tween
@@ -14,18 +13,12 @@ import androidx.compose.foundation.layout.systemBars
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.Dp
 import androidx.navigation3.runtime.entryProvider
 import androidx.navigation3.ui.NavDisplay
-import com.on.turip.R
 import com.on.turip.navigation.Navigator
 import com.on.turip.navigation.rememberNavigationState
 import com.on.turip.navigation.toEntries
@@ -50,39 +43,11 @@ fun MainApp(savedStateConfigurationProvider: SavedStateConfigurationProvider) {
 
     val appState = rememberDialogAppState(navigationState = navigationState)
     val navigator = remember { Navigator(appState.navigationState) }
-
-    var canExit: Boolean by rememberSaveable { mutableStateOf(false) }
-    val isTopLevelRoot: Boolean = appState.navigationState.isTopLevelKey
-
-    val exit: String = stringResource(R.string.main_double_back_pressed_to_exit)
+    ExitConfirmationHandler(appState = appState)
 
     val animatedSnackbarBottomPadding: Dp by animateDpAsState(
         targetValue = appState.snackbarBottomPadding,
         animationSpec = tween(durationMillis = 220),
-    )
-
-    LaunchedEffect(canExit, isTopLevelRoot) {
-        if (!canExit) return@LaunchedEffect
-
-        if (!isTopLevelRoot) {
-            canExit = false
-            appState.snackbarDelegate.dismissCurrentSnackbar()
-            return@LaunchedEffect
-        }
-
-        appState.snackbarDelegate.showSnackbar(
-            message = exit,
-            onDismiss = { canExit = false },
-        )
-    }
-
-    LaunchedEffect(appState.navigationState.currentTopLevelKey) {
-        appState.snackbarDelegate.dismissCurrentSnackbar()
-    }
-
-    BackHandler(
-        enabled = isTopLevelRoot && !canExit,
-        onBack = { canExit = true },
     )
 
     TuripTheme {
